Create Task 1
1. Fake News Detection Using Python
Fake news does not require any introduction. It is very much easy to spread
all the fake information in today’s all-connected world across the internet.
Fake news is sometimes transmitted through the internet by some
unauthorised sources, which creates issues for the targeted person and it
makes them panic and leads to even violence. To combat the spread of fake
news, it’s critical to determine the information’s legitimacy, which this Data
Science project can help with. To do so, Python can be used, and a model
is created using does. does can be implemented to distinguish between true
and fake news. Pandas, NumPy, and sci-kit-learn are some Python packages
suitable for this project, and we can utilize News.csv for the dataset.