Create Task 2
2.Detecting Forest Fire

Developing a project for identifying the forest fire and wildfire system
is an alternatively good example to exhibit one’s skills in Data Science.
The forest fire or wildfire is an uncontrollable fire that develops in a
forest. All the forest fir will create havoc during weekends on the
animal habitat, surrounding environment and human property. k-means
clustering can be used for the identification of the crucial hotspots
during forest fire and to reduce the severity, to regulate them and even
to predict the behaviour of the wildfire. This is advantageous for
allocating the required resources. To enhance the model’s accuracy, it
is ideal to use climatological data to find out the common periods and
seasons for wildfires.